+++ src/api/Models/Casing.cs
@@ -31,52 +31,6 @@ public class Casing : IChangeTracking, IIdentifyable
     public string Name { get; set; }
 
     /// <summary>
-    /// Gets or sets the <see cref="Casing"/>'s upper depth.
-    /// </summary>
-    [Column("from_depth")]
-    public double FromDepth { get; set; }
-
-    /// <summary>
-    /// Gets or sets the <see cref="Casing"/>'s lower depth.
-    /// </summary>
-    [Column("to_depth")]
-    public double ToDepth { get; set; }
-
-    /// <summary>
-    /// Gets or sets the <see cref="Casing"/>'s kind id.
-    /// </summary>
-    [Column("kind_id")]
-    public int KindId { get; set; }
-
-    /// <summary>
-    /// Gets or sets the <see cref="Casing"/>'s kind.
-    /// </summary>
-    public Codelist? Kind { get; set; }
-
-    /// <summary>
-    /// Gets or sets the id of the <see cref="Casing"/>'s material.
-    /// </summary>
-    [Column("material_id")]
-    public int MaterialId { get; set; }
-
-    /// <summary>
-    /// Gets or sets the <see cref="Casing"/>'s material.
-    /// </summary>
-    public Codelist? Material { get; set; }
-
-    /// <summary>
-    /// Gets or sets the <see cref="Casing"/>'s inner diameter.
-    /// </summary>
-    [Column("inner_diameter")]
-    public double InnerDiameter { get; set; }
-
-    /// <summary>
-    /// Gets or sets the <see cref="Casing"/>'s outer diameter.
-    /// </summary>
-    [Column("outer_diameter")]
-    public double OuterDiameter { get; set; }
-
-    /// <summary>
     /// Gets or sets the <see cref="Casing"/>'s start date.
     /// </summary>
     [Column("date_start")]
@@ -103,6 +57,11 @@ public class Casing : IChangeTracking, IIdentifyable
     /// Gets the <see cref="Observation"/>s associated with the <see cref="Casing"/>.
     /// </summary>
     public ICollection<Observation>? Observations { get; }
+
+    /// <summary>
+    /// Gets the <see cref="CasingElement"/>s of the <see cref="Casing"/>.
+    /// </summary>
+    public ICollection<CasingElement> CasingElements { get; set; }
 
     /// <inheritdoc />
     [Column("creator")]