+++ src/client/cypress/e2e/editor/backfill.cy.js
@@ -3,7 +3,7 @@ import { setInput, setSelect } from "../helpers/formHelpers";
 import { addItem, startEditing, saveForm, deleteItem } from "../helpers/buttonHelpers";
 
 describe("Backfill crud tests", () => {
-  it("add, edit and delete backfills", () => {
+  beforeEach(() => {
     createBorehole({ "extended.original_name": "INTEADAL" })
       .as("borehole_id")
       .then(id => createCompletion("test backfill", id, 16000002, true))
@@ -25,10 +25,11 @@ describe("Backfill crud tests", () => {
     // select backfill tab
     cy.get("[data-cy=completion-content-header-tab-backfill]").click();
     cy.wait("@backfill_GET");
+  });
 
+  it("add, edit and delete backfills", () => {
     // add new backfill card
     addItem("addFilling");
-    cy.get('[data-cy="addFilling-button"]').click({ force: true });
     cy.wait("@codelist_GET");
 
     // fill out form
@@ -62,5 +63,45 @@ describe("Backfill crud tests", () => {
     // delete backfill
     deleteItem();
     cy.contains("From depth").should("not.exist");
+  });
+
+  it("sort backfill", () => {
+    addItem("addFilling");
+    cy.wait("@codelist_GET");
+    setInput("notes", "Lorem.");
+    setInput("fromDepth", "0");
+    setInput("toDepth", "10");
+    setSelect("kindId", 2);
+    setSelect("materialId", 1);
+    saveForm();
+
+    cy.wait(1000);
+    addItem("addFilling");
+    cy.wait("@codelist_GET");
+    setInput("notes", "Lorem.");
+    setInput("fromDepth", "0");
+    setInput("toDepth", "12");
+    setSelect("kindId", 2);
+    setSelect("materialId", 1);
+    saveForm();
+
+    cy.get('[data-cy="backfill-card.0"] [data-cy="todepth-formDisplay"]').contains("10");
+    cy.get('[data-cy="backfill-card.1"] [data-cy="todepth-formDisplay"]').contains("12");
+
+    cy.get('[data-cy="backfill-card.1"] [data-cy="edit-button"]').click({
+      force: true,
+    });
+    setInput("toDepth", "8");
+    saveForm();
+    cy.get('[data-cy="backfill-card.0"] [data-cy="todepth-formDisplay"]').contains("8");
+    cy.get('[data-cy="backfill-card.1"] [data-cy="todepth-formDisplay"]').contains("10");
+
+    cy.get('[data-cy="backfill-card.0"] [data-cy="edit-button"]').click({
+      force: true,
+    });
+    setInput("fromDepth", "5");
+    saveForm();
+    cy.get('[data-cy="backfill-card.0"] [data-cy="fromdepth-formDisplay"]').contains("0");
+    cy.get('[data-cy="backfill-card.1"] [data-cy="fromdepth-formDisplay"]').contains("5");
   });
 });