+++ src/client/cypress/e2e/editor/instrumentation.cy.js
@@ -3,7 +3,7 @@ import { setInput, setSelect } from "../helpers/formHelpers";
 import { addItem, startEditing, saveForm, deleteItem } from "../helpers/buttonHelpers";
 
 describe("Instrumentation crud tests", () => {
-  it("add, edit and delete instrumentations", () => {
+  beforeEach(() => {
     createBorehole({ "extended.original_name": "INTEADAL" })
       .as("borehole_id")
       .then(id => createCompletion("test instr.", id, 16000002, true))
@@ -21,7 +21,9 @@ describe("Instrumentation crud tests", () => {
 
     // start editing session
     startBoreholeEditing();
+  });
 
+  it("add, edit and delete instrumentations", () => {
     // Precondition: Create casing to later link in instrumentation
     cy.get("[data-cy=completion-content-header-tab-casing]").click();
     cy.wait("@casing_GET");
@@ -30,15 +32,14 @@ describe("Instrumentation crud tests", () => {
     cy.wait("@codelist_GET");
 
     setInput("name", "casing-1");
-    setInput("fromDepth", "0");
-    setInput("toDepth", "10");
-    setSelect("kindId", 2);
-    setSelect("materialId", 3);
     setInput("dateStart", "2021-01-01");
     setInput("dateFinish", "2021-01-02");
-    setInput("innerDiameter", "3");
-    setInput("outerDiameter", "4");
-
+    setInput("casingElements.0.fromDepth", "0");
+    setInput("casingElements.0.toDepth", "10");
+    setSelect("casingElements.0.kindId", 2);
+    setSelect("casingElements.0.materialId", 3);
+    setInput("casingElements.0.innerDiameter", "3");
+    setInput("casingElements.0.outerDiameter", "4");
     saveForm();
     cy.wait("@casing_GET");
 
@@ -85,5 +86,53 @@ describe("Instrumentation crud tests", () => {
     // delete instrumentation
     deleteItem();
     cy.contains("From depth").should("not.exist");
+  });
+
+  it("sort instrumentation", () => {
+    cy.get("[data-cy=completion-content-header-tab-instrumentation]").click();
+    cy.wait("@instrumentation_GET");
+
+    addItem("addInstrument");
+    cy.wait("@casing_GET");
+    setInput("notes", "Lorem.");
+    setInput("name", "Inst-1");
+    setInput("fromDepth", "0");
+    setInput("toDepth", "10");
+    setSelect("kindId", 2);
+    setSelect("statusId", 1);
+    saveForm();
+    cy.wait("@instrumentation_GET");
+
+    addItem("addInstrument");
+    cy.wait("@casing_GET");
+    setInput("notes", "Lorem.");
+    setInput("name", "Inst-2");
+    setInput("fromDepth", "0");
+    setInput("toDepth", "12");
+    setSelect("kindId", 2);
+    setSelect("statusId", 1);
+    saveForm();
+    cy.wait("@instrumentation_GET");
+
+    cy.get('[data-cy="instrumentation-card.0"] [data-cy="name-formDisplay"]').contains("Inst-1");
+    cy.get('[data-cy="instrumentation-card.1"] [data-cy="name-formDisplay"]').contains("Inst-2");
+
+    cy.get('[data-cy="instrumentation-card.1"] [data-cy="edit-button"]').click({
+      force: true,
+    });
+    setInput("toDepth", "8");
+    saveForm();
+    cy.wait("@instrumentation_GET");
+    cy.get('[data-cy="instrumentation-card.0"] [data-cy="name-formDisplay"]').contains("Inst-2");
+    cy.get('[data-cy="instrumentation-card.1"] [data-cy="name-formDisplay"]').contains("Inst-1");
+
+    cy.get('[data-cy="instrumentation-card.0"] [data-cy="edit-button"]').click({
+      force: true,
+    });
+    setInput("fromDepth", "5");
+    saveForm();
+    cy.wait("@instrumentation_GET");
+    cy.get('[data-cy="instrumentation-card.0"] [data-cy="name-formDisplay"]').contains("Inst-1");
+    cy.get('[data-cy="instrumentation-card.1"] [data-cy="name-formDisplay"]').contains("Inst-2");
   });
 });