+++ src/client/src/commons/form/borehole/completion/backfill.jsx
@@ -3,6 +3,7 @@ import { getBackfills, addBackfill, updateBackfill, deleteBackfill } from "../..
 import { DataCards } from "../../../../components/dataCard/dataCards";
 import BackfillInput from "./backfillInput";
 import BackfillDisplay from "./backfillDisplay";
+import { sortByDepth } from "../../../sorter.jsx";
 
 const Backfill = ({ isEditable, completionId }) => {
   return (
@@ -13,10 +14,14 @@ const Backfill = ({ isEditable, completionId }) => {
       addData={addBackfill}
       updateData={updateBackfill}
       deleteData={deleteBackfill}
+      cyLabel="backfill"
       addLabel="addFilling"
       emptyLabel="msgFillingEmpty"
       renderInput={props => <BackfillInput {...props} />}
       renderDisplay={props => <BackfillDisplay {...props} />}
+      sortDisplayed={(a, b) => {
+        return sortByDepth(a, b, "fromDepth", "toDepth");
+      }}
     />
   );
 };