+++ src/client/src/commons/form/borehole/completion/instrumentation.jsx
@@ -8,6 +8,7 @@ import {
 import { DataCards } from "../../../../components/dataCard/dataCards";
 import InstrumentationInput from "./instrumentationInput";
 import InstrumentationDisplay from "./instrumentationDisplay";
+import { sortByDepth } from "../../../sorter.jsx";
 
 const Instrumentation = ({ isEditable, completionId }) => {
   return (
@@ -18,10 +19,14 @@ const Instrumentation = ({ isEditable, completionId }) => {
       addData={addInstrumentation}
       updateData={updateInstrumentation}
       deleteData={deleteInstrumentation}
+      cyLabel="instrumentation"
       addLabel="addInstrument"
       emptyLabel="msgInstrumentsEmpty"
       renderInput={props => <InstrumentationInput {...props} />}
       renderDisplay={props => <InstrumentationDisplay {...props} />}
+      sortDisplayed={(a, b) => {
+        return sortByDepth(a, b, "fromDepth", "toDepth");
+      }}
     />
   );
 };