+++ src/client/src/commons/form/borehole/hydrogeology/groundwaterLevelMeasurement.jsx
@@ -7,6 +7,7 @@ import {
 import GroundwaterLevelMeasurementInput from "./groundwaterLevelMeasurementInput";
 import GroundwaterLevelMeasurementDisplay from "./groundwaterLevelMeasurementDisplay";
 import { DataCards } from "../../../../components/dataCard/dataCards";
+import { sortByDepth } from "../../../sorter.jsx";
 
 const GroundwaterLevelMeasurement = ({ isEditable, boreholeId }) => {
   return (
@@ -17,10 +18,14 @@ const GroundwaterLevelMeasurement = ({ isEditable, boreholeId }) => {
       addData={addGroundwaterLevelMeasurement}
       updateData={updateGroundwaterLevelMeasurement}
       deleteData={deleteGroundwaterLevelMeasurement}
+      cyLabel="groundwaterLevelMeasurement"
       addLabel="addGroundwaterLevelMeasurement"
       emptyLabel="msgGroundwaterLevelMeasurementsEmpty"
       renderInput={props => <GroundwaterLevelMeasurementInput {...props} />}
       renderDisplay={props => <GroundwaterLevelMeasurementDisplay {...props} />}
+      sortDisplayed={(a, b) => {
+        return sortByDepth(a, b, "fromDepthM", "toDepthM");
+      }}
     />
   );
 };