+++ src/client/src/commons/form/borehole/hydrogeology/hydrotest.jsx
@@ -2,6 +2,7 @@ import { getHydrotests, addHydrotest, updateHydrotest, deleteHydrotest } from ".
 import HydrotestInput from "./hydrotestInput";
 import HydrotestDisplay from "./hydrotestDisplay";
 import { DataCards } from "../../../../components/dataCard/dataCards";
+import { sortByDepth } from "../../../sorter.jsx";
 
 const Hydrotest = ({ isEditable, boreholeId }) => {
   return (
@@ -12,10 +13,14 @@ const Hydrotest = ({ isEditable, boreholeId }) => {
       addData={addHydrotest}
       updateData={updateHydrotest}
       deleteData={deleteHydrotest}
+      cyLabel="hydrotest"
       addLabel="addHydrotest"
       emptyLabel="msgHydrotestEmpty"
       renderInput={props => <HydrotestInput {...props} />}
       renderDisplay={props => <HydrotestDisplay {...props} />}
+      sortDisplayed={(a, b) => {
+        return sortByDepth(a, b, "fromDepthM", "toDepthM");
+      }}
     />
   );
 };