+++ src/client/src/commons/form/borehole/hydrogeology/waterIngress.jsx
@@ -2,6 +2,7 @@ import { getWaterIngress, addWaterIngress, updateWaterIngress, deleteWaterIngres
 import WaterIngressInput from "./waterIngressInput";
 import WaterIngressDisplay from "./waterIngressDisplay";
 import { DataCards } from "../../../../components/dataCard/dataCards";
+import { sortByDepth } from "../../../sorter.jsx";
 
 const WaterIngress = ({ isEditable, boreholeId }) => {
   return (
@@ -12,10 +13,14 @@ const WaterIngress = ({ isEditable, boreholeId }) => {
       addData={addWaterIngress}
       updateData={updateWaterIngress}
       deleteData={deleteWaterIngress}
+      cyLabel="waterIngress"
       addLabel="addWaterIngress"
       emptyLabel="msgWateringressesEmpty"
       renderInput={props => <WaterIngressInput {...props} />}
       renderDisplay={props => <WaterIngressDisplay {...props} />}
+      sortDisplayed={(a, b) => {
+        return sortByDepth(a, b, "fromDepthM", "toDepthM");
+      }}
     />
   );
 };