+++ src/client/src/commons/form/borehole/hydrogeology/waterIngressInput.jsx
@@ -20,6 +20,7 @@ const WaterIngressInput = props => {
     if (data.casingId == null) {
       data.casingId = item.casingId;
     }
+    data.casing = null;
     if (data.conditionsId === "") {
       data.conditionsId = null;
     }