+++ tests/Controllers/CasingControllerTest.cs
@@ -1,5 +1,6 @@
 ﻿using BDMS.Models;
 using Microsoft.AspNetCore.Http;
+using Microsoft.CodeAnalysis.Emit;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -63,51 +64,78 @@ public class CasingControllerTest
     }
 
     [TestMethod]
+    public async Task CreateWithoutElementAsync()
+    {
+        var casing = new Casing()
+        {
+            CompletionId = CompletionId,
+            Name = "COLLAR",
+            DateStart = new DateOnly(2021, 1, 1),
+            DateFinish = new DateOnly(2021, 1, 2),
+            Notes = "ARGONSHIP",
+        };
+
+        var response = await controller.CreateAsync(casing);
+        ActionResultAssert.IsInternalServerError(response.Result, "At least one casing element must be defined.");
+    }
+
+    [TestMethod]
     public async Task CreateAsync()
     {
         var casing = new Casing()
         {
             CompletionId = CompletionId,
             Name = "COLLAR",
-            FromDepth = 0,
-            ToDepth = 100,
-            MaterialId = context.Codelists.First(c => c.Schema == CompletionSchemas.CasingMaterialSchema).Id,
-            KindId = context.Codelists.First(c => c.Schema == CompletionSchemas.CasingTypeSchema).Id,
             DateStart = new DateOnly(2021, 1, 1),
             DateFinish = new DateOnly(2021, 1, 2),
-            InnerDiameter = 3,
-            OuterDiameter = 4,
             Notes = "ARGONSHIP",
+            CasingElements = new List<CasingElement>()
+            {
+                new CasingElement()
+                {
+                    FromDepth = 0,
+                    ToDepth = 100,
+                    MaterialId = context.Codelists.First(c => c.Schema == CompletionSchemas.CasingMaterialSchema).Id,
+                    KindId = context.Codelists.First(c => c.Schema == CompletionSchemas.CasingTypeSchema).Id,
+                    InnerDiameter = 3,
+                    OuterDiameter = 4,
+                },
+            },
         };
 
         var response = await controller.CreateAsync(casing);
         ActionResultAssert.IsOkObjectResult<Casing>(response.Result);
 
-        casing = await context.Casings.FindAsync(casing.Id);
+        casing = await context.Casings
+            .Include(c => c.CasingElements)
+            .AsNoTracking()
+            .SingleOrDefaultAsync(c => c.Id == casing.Id);
+
         Assert.IsNotNull(casing);
         Assert.AreEqual(CompletionId, casing.CompletionId);
         Assert.AreEqual("ARGONSHIP", casing.Notes);
-        Assert.AreEqual(0, casing.FromDepth);
-        Assert.AreEqual(100, casing.ToDepth);
-        Assert.AreEqual(context.Codelists.First(c => c.Schema == CompletionSchemas.CasingMaterialSchema).Id, casing.MaterialId);
-        Assert.AreEqual(context.Codelists.First(c => c.Schema == CompletionSchemas.CasingTypeSchema).Id, casing.KindId);
-        Assert.AreEqual(3, casing.InnerDiameter);
-        Assert.AreEqual(4, casing.OuterDiameter);
         Assert.AreEqual(new DateOnly(2021, 1, 1), casing.DateStart);
         Assert.AreEqual(new DateOnly(2021, 1, 2), casing.DateFinish);
+        Assert.AreEqual(1, casing.CasingElements.Count);
+        var casingElement = casing.CasingElements.First();
+        Assert.AreEqual(0, casingElement.FromDepth);
+        Assert.AreEqual(100, casingElement.ToDepth);
+        Assert.AreEqual(context.Codelists.First(c => c.Schema == CompletionSchemas.CasingMaterialSchema).Id, casingElement.MaterialId);
+        Assert.AreEqual(context.Codelists.First(c => c.Schema == CompletionSchemas.CasingTypeSchema).Id, casingElement.KindId);
+        Assert.AreEqual(3, casingElement.InnerDiameter);
+        Assert.AreEqual(4, casingElement.OuterDiameter);
     }
 
     [TestMethod]
     public async Task EditAsync()
     {
-        var casing = context.Casings.First();
+        var casing = await context.Casings
+            .Include(c => c.CasingElements)
+            .AsNoTracking()
+            .FirstAsync(c => c.CasingElements.Count > 0);
         var completionId = casing.CompletionId;
 
-        casing.MaterialId = context.Codelists.First(c => c.Schema == CompletionSchemas.CasingMaterialSchema).Id;
-        casing.KindId = context.Codelists.First(c => c.Schema == CompletionSchemas.CasingTypeSchema).Id;
         casing.Notes = "COLLAR";
-        casing.FromDepth = 50;
-        casing.ToDepth = 200;
 
         var response = await controller.EditAsync(casing);
         ActionResultAssert.IsOkObjectResult<Casing>(response.Result);
@@ -116,10 +144,6 @@ public class CasingControllerTest
         Assert.IsNotNull(casing);
         Assert.AreEqual(completionId, casing.CompletionId);
         Assert.AreEqual("COLLAR", casing.Notes);
-        Assert.AreEqual(50, casing.FromDepth);
-        Assert.AreEqual(200, casing.ToDepth);
-        Assert.AreEqual(context.Codelists.First(c => c.Schema == CompletionSchemas.CasingMaterialSchema).Id, casing.MaterialId);
-        Assert.AreEqual(context.Codelists.First(c => c.Schema == CompletionSchemas.CasingTypeSchema).Id, casing.KindId);
     }
 
     [TestMethod]